fix hide after outside click
fix close popin ok

fix style of items card

fix style after rebase
