+++ packages/@coorpacademy-components/src/molecule/cm-popin/index.js
@@ -1,4 +1,5 @@
-import React from 'react';
+// import React from 'react';
+import React, {useRef, useEffect, useState} from 'react';
 import {
   NovaSolidInterfaceFeedbackInterfaceAlertDiamond as AlertDiamond,
   NovaSolidApplicationsWindowUpload3 as WindowUpload,
@@ -60,7 +61,21 @@ const CMPopin = props => {
         backgroundSize: 'cover'
       }
     : null;
-
+  const [shloudRender, setShouldRendered] = useState(true);
+  const nodeRef = useRef(null);
+  useEffect(() => {
+    if (mode === 'items') {
+      const closePopin = e => {
+        if (nodeRef && nodeRef.current && !nodeRef.current.contains(e.target)) {
+          setShouldRendered(false);
+        }
+      };
+      document.addEventListener('click', closePopin);
+      return () => {
+        document.removeEventListener('click', closePopin);
+      };
+    }
+  }, [mode]);
   const renderHeader = () => {
     if (header) {
       const {title, headerIcon, backgroundImage} = header;
@@ -165,16 +180,16 @@ const CMPopin = props => {
   const wrapperPopinStyle = classNames(
     mode === 'cookie' && style.popinCookie,
     mode === 'list' && style.popinFilesList,
+    mode === 'items' && style.popinItems,
     style.popin
   );
-
-  return (
+  return shloudRender ? (
     <div
       className={mode !== 'cookie' ? style.background : null}
       style={backgroundImageStyle}
       data-name={'cm-popin-container'}
     >
-      <div className={wrapperPopinStyle}>
+      <div className={wrapperPopinStyle} ref={nodeRef}>
         <header className={style.popinHeader}>
           {renderHeader()}
           {onClose ? (
@@ -260,7 +275,7 @@ const CMPopin = props => {
         ) : null}
       </div>
     </div>
-  );
+  ) : null;
 };
 
 CMPopin.propTypes = propTypes;
