+++ packages/@coorpacademy-components/src/template/common/dashboard/style.css
@@ -11,10 +11,10 @@
   height: 350px;
   width: 100%;
 }
-
+/* 
 .hidden {
   display: none;
-}
+} */
 
 .popinWithCards {
   position: relative;
