+++ packages/@coorpacademy-components/src/template/common/search-page/index.js
@@ -1,4 +1,4 @@
-import React, {useCallback, useEffect, useRef} from 'react';
+import React, {useRef} from 'react';
 import PropTypes from 'prop-types';
 import {getOr, isEmpty} from 'lodash/fp';
 import Provider from '../../../atom/provider';
@@ -27,17 +27,6 @@ const SearchPage = (props, context) => {
   const {skin} = context;
   const defaultColor = getOr('#00B0FF', 'common.primary', skin);
   const nodeRef = useRef(null);
-  const closePopin = useCallback(e => {
-    if (nodeRef && nodeRef.current && nodeRef.current.contains(e.target)) {
-      e.target.className = style.hidden;
-    }
-  }, []);
-  useEffect(() => {
-    document.addEventListener('click', closePopin);
-    return () => {
-      document.removeEventListener('click', closePopin);
-    };
-  }, [closePopin]);
   const recommendationsView = isEmpty(recommendations) ? null : <CardsList {...recommendations} />;
 
   const cardsView = isEmpty(cards.list) ? (
@@ -59,7 +48,7 @@ const SearchPage = (props, context) => {
   );
 
   return (
-    <div ref={nodeRef}>
+    <div>
       <Filters
         {...searchFilters}
         moreSortAriaLabel={moreSortAriaLabel}
@@ -74,7 +63,7 @@ const SearchPage = (props, context) => {
         {cardsView}
       </div>
       {popinWithCards ? (
-        <div className={style.popinWithCards}>
+        <div className={style.popinWithCards} ref={nodeRef}>
           <CMPopin {...popinWithCards} />
         </div>
       ) : null}
