+++ packages/@coorpacademy-components/src/template/common/search-page/style.css
@@ -17,9 +17,9 @@
   background: white;
 }
 
-.hidden {
+/* .hidden {
   display: none;
-}
+} */
 
 .noresults {
   font-family: "Gilroy";
